backend/api/pyproject.toml
[tool.poetry]
name = "bee-to-bee-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^1.1.1"
flask-sqlalchemy = "^2.4.1"
Flask-GraphQL = "^2.0.1"
graphene-sqlalchemy = "^2.2.2"
graphene = "^2.1.8"
psycopg2-binary = "^2.8.4"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pylint = "^2.4.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
